requirements.txt
fastapi
uvicorn
gunicorn
python-dotenv
pydantic
requests
google-generativeai
